Range Locking: Release the iterator when taking a lock on a range
This will cause the iterator to be refreshed so that it returns
the latest committed data, and not the data from a snapshot taken
at the point the iterator was created.

Iterator::Refresh() is not usable for us because this method is
not implemented in rocksdb::BaseDeltaIterator that
Transaction::GetIterator() returns.