Apply patch: Merge index scan with index reads
Summary:
This merges index scans with index lookups. This is so that the codepaths to perform retry with snapshot refresh are consolidated into one location.

Also, in `position_to_correct_key`, I save a redundant call to `pack_index_tuple` since we just packed it earlier. To do this, the packed tuple is saved in `m_sk_match_prefix_buf`.

Test Plan: mtr

Reviewers: luqun, herman, yzha

Subscribers: pgl, vinaybhat

Differential Revision: https://phabricator.intern.facebook.com/D23358423