Apply patch: Support parsing index comments for partial indexes
Summary:
This adds support for parsing the comment section on secondary keys for partial index keywords. This is then populated onto the `Rdb_key_def` structure.

This isn't persisted to the data dictionary because we don't really need this information for compaction (unlike some of our TTL related fields).

Test Plan: mtr

Reviewers: luqun, herman, yzha, #mysql_eng

Subscribers: pgl

Differential Revision: https://phabricator.intern.facebook.com/D25933175