Apply patch: Add partial index iterator
Summary:
This adds the partial index iterator. It is a special iterator that sorts groups from the primary key on the fly as needed, and if it exceeds a certain threshold, it will materialize the rows on the secondary key as well.

For point queries, the secondary key read is not found, we simply extract the primary key parts and read from the primary key. This means that point queries don't trigger materialization though.

Test Plan: mtr

Reviewers: luqun, herman, yzha, #mysql_eng

Subscribers: phabricatorlinter, pgl

Differential Revision: https://phabricator.intern.facebook.com/D25933178